+++ makefile
@@ -6,3 +6,6 @@ export_poetry:
 
 auto_commit:
 	git add . && git commit -m 'Update' && git push
+
+run_app:
+	poetry run streamlit run myapp.py
