+++ pyproject.toml
@@ -18,6 +18,7 @@ yfinance = "^0.2.12"
 pandas-datareader = "^0.10.0"
 scipy = "^1.10.1"
 pre-commit = "^3.5.0"
+loguru = "^0.7.2"
 
 
 [tool.poetry.group.dev.dependencies]
